W4 warnings
ahhh, now it feels like home. I'm used to having -W4 -WX on lately for my other project. Idk I like the compiler helping you like this.

+++ code/base/vn_math.cpp
@@ -947,7 +947,7 @@ f32 MaxF32(f32 i, f32 j)
 	return result;
 }
 
-f32 ClampI32(i32 min, i32 value, i32 max)
+i32 ClampI32(i32 min, i32 value, i32 max)
 {
 	i32 result = value;
